refactor(routes): move gallery domain into routes/gallery/ subpackage
Slice 2a of the route-domain reorganization (#4082/#4071, per
specs/architecture-runtime-inventory.md §6.3). Moves gallery_routes.py and
gallery_helpers.py into routes/gallery/, leaving backward-compat re-export
shims at the old paths. Pure file reorganization, no behavior change.

The shims use sys.modules replacement so that `import routes.gallery_routes`,
`from routes.gallery_routes import X`, `importlib.import_module(...)`, and
`monkeypatch.setattr(routes.gallery_routes, ...)` all operate on the same
module object the application uses — a plain `import *` shim drops private
names and breaks monkeypatch-based tests.

9 source-introspection test sites (Path().read_text) and 2 importlib sites
are repointed to the new canonical path; email entries in the shared
test_upload_limits_centralized / test_direct_upload_limits dicts are left
untouched.

Verified: compileall clean; full suite 4054 passed, 2 skipped.
